Rename llvm datasets to legacy datasets.
github.com/facebookresearch/CompilerGym/issues/45
