+++ tests/fuzzing/llvm_validate_fuzz_test.py
@@ -6,7 +6,9 @@
 import random
 
 from compiler_gym.envs import LlvmEnv
-from compiler_gym.envs.llvm.datasets import get_llvm_benchmark_validation_callback
+from compiler_gym.envs.llvm.legacy_datasets import (
+    get_llvm_benchmark_validation_callback,
+)
 from tests.pytest_plugins.llvm import VALIDATABLE_BENCHMARKS
 from tests.test_main import main
